requirements.txt
apache-airflow~=2.10.4
psycopg2-binary~=2.9.10
Faker~=33.3.0